wasteless-bites-api/requirements.txt

# jwtdown-fastapi>=0.2.0
fastapi==0.78.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[brcypt]>=1.7.4,<2.0.0
uvicorn[standard]>=0.17.6,<0.18.0
python-multipart==0.0.5
aiofiles>=0.8.0,<0.9.0
psycopg[binary,pool]==3.1.2
pyjwt>=2.6.0
